[refactor] Move all format-related functionality into `logform`.
