+++ lib/winston.js
@@ -27,8 +27,13 @@ winston.transports = require('./winston/transports');
 // Expose utility methods
 //
 winston.config    = require('./winston/config');
-winston.addColors = winston.config.addColors;
-winston.format    = require('./winston/formats');
+
+//
+// Hoist format-related functionality from logform.
+//
+const logform = require('logform');
+winston.addColors = logform.levels;
+winston.format    = logform.format;
 
 //
 // Expose core Logging-related prototypes.
