--- lib/winston/formats/pad-levels.js
@@ -1,13 +1,0 @@
-'use strict';
-
-var format = require('./format');
-
-/*
- * function padLevels (opts)
- * Returns a new instance of the padLevels format TransformStream
- * which pads levels to be the same length. This was previously
- * exposed as the `padLevels` option to transports in `winston < 3.0.0`.
- */
-module.exports = format(function (info, opts) {
-
-});
