+++ lib/winston/logger.js
@@ -48,7 +48,7 @@ Logger.prototype.configure = function (options) {
   // and then adds another TransportStream?
   //
   options = options || {};
-  this.format = options.format || this.format || require('./formats/json')();
+  this.format = options.format || this.format || require('logform/json')();
 
   var levels = options.levels || this.levels || config.npm.levels;
   var maxLength = Math.max.apply(null, Object.keys(levels)
