+++ test/fixtures/file/abstract.log
@@ -37,3 +37,6 @@
 {"level":"debug","message":"foo"}
 {"level":"debug","message":"foo"}
 {"level":"debug","message":"foo"}
+{"level":"debug","message":"foo"}
+{"level":"debug","message":"foo"}
+{"level":"debug","message":"foo"}
