+++ test/fixtures/file/simple.log
@@ -20,3 +20,4 @@ this is my log message
 this is my log message
 this is my log message
 this is my log message
+this is my log message
